#38 fixes to predict
